requirements.txt
requests==2.32.3
beautifulsoup4==4.12.2
python-memcached==1.59
redis==5.0.1
pybloom-live==2.3.1
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
uvloop==0.19.0
